requirements.txt
fastapi==0.115.0
httpx==0.27.2
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pytest==8.3.3
python-jose==3.3.0
sqlalchemy==2.0.35
uvicorn==0.30.6
